move test structs to test_utils
