+++ block-producer/src/state_view/tests/apply_block.rs
@@ -6,6 +6,8 @@
 
 use std::iter;
 
+use crate::test_utils::MockStoreSuccess;
+
 use super::*;
 
 /// Tests requirement AB1
