+++ block-producer/src/state_view/tests/verify_tx.rs
@@ -14,6 +14,8 @@ use std::iter;
 
 use tokio::task::JoinSet;
 
+use crate::test_utils::MockStoreSuccess;
+
 use super::*;
 
 /// Tests the happy path where 3 transactions who modify different accounts and consume different
