+++ block-producer/src/test_utils/mod.rs
@@ -1,2 +1,14 @@
+use std::collections::{BTreeMap, BTreeSet};
+use std::sync::Arc;
+use tokio::sync::RwLock;
+
+use miden_objects::{accounts::AccountId, Digest};
+
 mod proven_tx;
 pub use proven_tx::DummyProvenTxGenerator;
+
+mod store;
+pub use store::{MockStoreFailure, MockStoreSuccess};
+
+mod account;
+pub use account::MockPrivateAccount;
